requirements-lock.txt
anyascii==0.3.3
audioread==3.0.1
boto3==1.40.42
botocore==1.40.42
cached_path==1.8.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
docopt @ git+https://github.com/docopt/docopt.git@944641a7357376a43151bd4a3ca44f9c07c4d1ae
filelock==3.13.1
fsspec==2024.6.1
g2pkk==0.1.2
google-api-core==2.25.1
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
inflect==7.5.0
jamo==0.4.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.3
nltk==3.9.1
num2words==0.5.12
numba==0.62.1
numpy==2.3.3
packaging==25.0
pandas==2.3.2
platformdirs==4.4.0
pooch==1.8.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
resampy==0.4.3
rich==13.9.4
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
tqdm==4.67.1
transformers==4.56.2
txtsplit @ git+https://github.com/fakerybakery/txtsplit.git@fb444d4e6f01a136c0369bc01452082abf0f0d3c
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
wheel==0.45.1
